Fix devices getting unselected when linking to Google Home
When linking to Google Home, Hubitat attempts to update
all device properties with the devices selected during
the OAuth flow.  This doesn't work right if the property
name is a GString, which it will be by default if
interpolation is used.  Explicitly convert it to a String
to work around this issue.