Add support for using a single command in the OnOff trait
This adds the ability to configure the OnOff trait to use a single
command that accepts different parameters to turn its device on or
off.

It also changes the on and off attributes so that only one is
required.

These two changes make it possible to control ceiling fans purely
via the interface provided by the fanControl capability