Add confirmation and PIN code support
This adds support for configuring some actions of a defice type
to require either a confirmation or a PIN code.  The former is
mainly useful to prevent accidentally triggering an action.  The
latter is intended to secure devices such as doors and locks to
prevent their actions from being triggered by unauthorized parties.