Implement the queryOnlyTemperatureSetting attribute
This adds support for standalone ambient temperature sensors
using the queryOnlyTemperatureSetting attribute of the
TemperatureSetting trait.  It's not perfect, but this allows you
to set up ambient temperature sensors and query their temperature
by asking the Google Assistant things like "What's the temperature
of {device}?" or "What's the temperature in {room}?".  When you
ask for the temperature of a room the assistant will respond with
the temperature of each temperature sensor in the room
individually.