Added Check for deviceInfo Being Null (#33)
I was having a problem every time Google would to to sync devices and something wasn't found.  For the life of me I couldn't figure out what had happened.  Also, the synce would never finish completely because It was erroring out.  The additional couple lines of code allowed me to track down the offending dvice and fix it.  But even before I was able to do that, it allowed the sync to finish successfully rather than erroring out.