Handle command execution errors better
This works around the hacky way I'm working around the fact that
Hubitat doesn't let you define your own exception classes.
If command execution throws an exception, we now first try to parse
that exception's messag as JSON to generate a specific error response.
If that fails, we log the error and report a hard error back to
Google.