Avoid breaking the whole app because of a bad toggle (#55)
When creating a new toggle in the Toggles trait it's possible to
leave the page without filling in any labels by either closing the
browser or clicking a link out of the app.  This would break the
entire app settings due to an error trying to load that bad toggle.

This adds a default "Unknown" label that will be used in that case
so that the app will continue to function and the user can delete
the bad toggle.