Add support for new Hubitat rooms feature (#70)
This adds support for the new rooms feature that was added in Hubita
version 2.2.7.  This app will now populate the roomHint field on devices
based on their Hubitat room name when syncing with Google Home.  This
allows new devices to automatically be placed in the correct room in
Google Home.