Log a warning if a device is selected as multiple device types (#82)
The app currently doesn't try to reconcile when a device is selected as
multiple device types.  Previously, it would just send two device
configurations to Google with the same ID, causing Google to reject the
SYNC response outright.  This changes the behavior in that situation to
log a warning and ignore subsequent definitions of the same device so
that one misconfigured device doesn't block the user's other devices
from being synced with Google.