Uppercase values sent for MediaState attributes (#104)
This PR includes a tweak to the error reporting when posting device state requests so that failures will also output the request body that caused the error. I found it easier to include in the error message as there's less content in the logs to dig through to identify the request that failed. No problem if you want me to remove this as I see that debug logging can be enabled to output the same info if needed.