Fix hub version comparison (#105)
The hubVersionLessThan function would incorrectly return true if the hub
version was greater than the target version but contained a piece that
was less than the corresponding piece in the target version.  For
example, 2.3.5.103 would be considered less than 2.3.4.115 because 103
is less than 115 even though comparison should have stopped after
comparing 5 to 4.