Fix setVolume command for devices whose volume can't be set directly
The Google Home app always shows the volume slider and uses setVolume
commands for devices that repor their volume.  That caused devices
without a direct setVolume command not work correctly with the app.
This updates the setVolume handler to incremntally adjust the volume one
step at a time until it reaches the desired volume.