Fix CodeNarc and enable new rules (#114)
The CodeNarc checks were crashing because CodeNarc removed one of their
rule classes.  This removes that rule from the config and adds the new
rules that have been added since I last updated it.  It also fixes all
of the issues that have crept in since the checks haven't been running.